+++ packages/playground/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @internal/playground
 
+## 1.20.0-alpha.9
+
+### Patch Changes
+
+- Updated dependencies [[`0a2c22c`](https://github.com/mastra-ai/mastra/commit/0a2c22c902604439ec490319e14c17f331e0c84c), [`3a8024c`](https://github.com/mastra-ai/mastra/commit/3a8024ce615f8aa89479c0d71fe61d10bb0040be)]:
+  - @mastra/core@1.52.0-alpha.9
+  - @mastra/client-js@1.33.0-alpha.9
+  - @mastra/react@1.2.6-alpha.9
+  - @mastra/playground-ui@42.0.0-alpha.9
+
+## 1.20.0-alpha.8
+
+### Patch Changes
+
+- Updated dependencies [[`a4620ca`](https://github.com/mastra-ai/mastra/commit/a4620cae0421e880af53e9a4fa324a3e0f079a17), [`3b77e77`](https://github.com/mastra-ai/mastra/commit/3b77e7704936522e4769d29de1b5ea6901f302bd), [`6b1bf3b`](https://github.com/mastra-ai/mastra/commit/6b1bf3b9494bd51aa8f654c68c9355d6046fa2a1), [`72e437c`](https://github.com/mastra-ai/mastra/commit/72e437c515942c80b9def5b026e0bdee61b469d9)]:
+  - @mastra/playground-ui@42.0.0-alpha.8
+  - @mastra/core@1.52.0-alpha.8
+  - @mastra/client-js@1.33.0-alpha.8
+  - @mastra/react@1.2.6-alpha.8
+
 ## 1.20.0-alpha.7
 
 ### Patch Changes